feat: Added ABC for Database module
